Address PR comments: consolidate KMS validation & reduce debug logging
1. Create shared s3_validation_utils.go for consistent KMS key validation
   - Move isValidKMSKeyID from s3_sse_kms.go to shared utility
   - Ensures consistent validation across bucket encryption, object operations, and copy validation
   - Eliminates coupling between s3_bucket_encryption.go and s3_sse_kms.go
   - Provides comprehensive validation: rejects spaces, control characters, validates length

2. Reduce verbose debug logging in calculateIVWithOffset function
   - Change glog.Infof to glog.V(4).Infof for debug statements
   - Prevents log flooding in production environments
   - Consistent with other debug logs in the codebase

Both changes improve code quality, maintainability, and production readiness.
